# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MoneyMind AI - Personal Finance Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/trends:
    get:
      operationId: getAnalyticsTrends
      tags: [analytics]
      summary: Get income/expense/savings trend data for charts
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [weekly, monthly, yearly]
        - name: months
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendsData"
  /analytics/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [analytics]
      summary: Get spending breakdown by category
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [expense, income]
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdownItem"
  /analytics/heatmap:
    get:
      operationId: getExpenseHeatmap
      tags: [analytics]
      summary: Get daily expense heatmap data
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeatmapDay"
...
  /analytics/report/weekly:
    get:
      operationId: getWeeklyReport
      tags: [analytics]
      summary: Get weekly financial report
      parameters:
        - name: week_start
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Weekly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReport"
  /analytics/report/monthly:
    get:
      operationId: getMonthlyReport
      tags: [analytics]
      summary: Get monthly financial report
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
  /analytics/cashflow:
    get:
      operationId: getCashFlow
      tags: [analytics]
      summary: Get daily cashflow data for a date range
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cashflow data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashFlowDay"
components:
  schemas:
...
    PeriodSummary:
      type: object
      required: [income, expenses, savings]
      properties:
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number

    # ── Analytics ─────────────────────────────────────────────────────────────
    TrendsData:
      type: object
      required: [labels, income, expenses, savings]
      properties:
        labels:
          type: array
          items:
            type: string
        income:
          type: array
          items:
            type: number
        expenses:
          type: array
          items:
            type: number
        savings:
          type: array
          items:
            type: number
    CategoryBreakdownItem:
      type: object
      required: [category_id, category_name, amount, percent]
      properties:
        category_id:
          type: integer
        category_name:
          type: string
        category_color:
          type: ["string", "null"]
        category_icon:
          type: ["string", "null"]
        amount:
          type: number
        percent:
          type: number
        transaction_count:
          type: integer
    HeatmapDay:
      type: object
      required: [date, amount]
      properties:
        date:
          type: string
        amount:
          type: number
        level:
          type: integer
...
    WeeklyReport:
      type: object
      required: [week_start, week_end, summary]
      properties:
        week_start:
          type: string
        week_end:
          type: string
        summary:
          $ref: "#/components/schemas/PeriodSummary"
        top_categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdownItem"
        biggest_expense_day:
          type: ["string", "null"]
        lowest_expense_day:
          type: ["string", "null"]
        best_saving_day:
          type: ["string", "null"]
        daily_data:
          type: array
          items:
            $ref: "#/components/schemas/CashFlowDay"
        ai_summary:
          type: ["string", "null"]
    MonthlyReport:
      type: object
      required: [year, month, summary]
      properties:
        year:
          type: integer
        month:
          type: integer
        summary:
          $ref: "#/components/schemas/PeriodSummary"
        top_categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdownItem"
        budget_accuracy:
          type: number
        savings_rate:
          type: number
        previous_month_comparison:
          $ref: "#/components/schemas/PeriodSummary"
        highest_expense:
          type: ["number", "null"]
        lowest_expense:
          type: ["number", "null"]
        daily_data:
          type: array
          items:
            $ref: "#/components/schemas/CashFlowDay"
        ai_summary:
          type: ["string", "null"]
    CashFlowDay:
      type: object
      required: [date, income, expenses, savings]
      properties:
        date:
          type: string
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number
        balance:
          type: number