# lib/api-zod/src/generated/types/categoryBreakdownItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MoneyMind AI - Personal Finance Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryBreakdownItem {
  category_id: number;
  category_name: string;
  /** @nullable */
  category_color?: string | null;
  /** @nullable */
  category_icon?: string | null;
  amount: number;
  percent: number;
  transaction_count?: number;
}