# lib/api-zod/src/generated/types/weeklyReport.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MoneyMind AI - Personal Finance Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { CashFlowDay } from './cashFlowDay';
import type { CategoryBreakdownItem } from './categoryBreakdownItem';
import type { PeriodSummary } from './periodSummary';

export interface WeeklyReport {
  week_start: string;
  week_end: string;
  summary: PeriodSummary;
  top_categories?: CategoryBreakdownItem[];
  /** @nullable */
  biggest_expense_day?: string | null;
  /** @nullable */
  lowest_expense_day?: string | null;
  /** @nullable */
  best_saving_day?: string | null;
  daily_data?: CashFlowDay[];
  /** @nullable */
  ai_summary?: string | null;
}